Move ApplyMask into cpp file
