+++ query/pipeline/processors/ApplyMask.h
@@ -1,28 +1,6 @@
 #pragma once
 
-#include "columns/AllowedKinds.h"
-#include "columns/ColumnCombinations.h"
-#include "columns/ColumnOperatorDispatcher.h"
-#include "columns/MaskOperators.h"
-
 namespace db {
-
-namespace {
-
-struct Eval {
-    Column* _res {nullptr};
-    
-    template <typename T, typename U>
-    void operator()(const T* arg, const U* mask) {
-        bioassert(_res && arg && mask, "Invalid operands to evaluate mask");
-
-        using ResultType = ColumnCombination<Apply, T, U>::ResultColumnType;
-        auto* result = dynamic_cast<ResultType*>(_res);
-        MaskApplicator::apply(result, arg, mask);
-    }
-};
-
-}
 
 class Column;
 
@@ -30,15 +8,5 @@ class ApplyMask {
 public:
     static void eval(Column* res, const Column* arg, const Column* mask);
 };
-
-inline void ApplyMask::eval(Column* res, const Column* arg, const Column* mask) {
-    using Pairs = MaskedPairs;
-    using Dispatcher = ColumnDoubleDispatcher<Pairs::Allowed,
-                                              Pairs::AllowedMixed,
-                                              Eval,
-                                              Pairs::Excluded>;
-    Eval fn {res};
-    Dispatcher::dispatch(arg, mask, fn);
-}
 
 }
