+++ query/pipeline/processors/FilterProcessor.h
@@ -11,7 +11,7 @@ class PredicateProgram;
 
 class FilterProcessor final : public Processor {
 public:
-    static FilterProcessor* create(PipelineV2* pipeline, PredicateProgram* predProg);
+    static FilterProcessor* create(PipelineV2* pipeline, PredicateProgram* prog);
 
     void prepare(ExecutionContext* ctxt) final;
     void reset() final;
@@ -26,9 +26,9 @@ private:
     PipelineBlockInputInterface _input;
     PipelineBlockOutputInterface _output;
 
-    PredicateProgram* _predProg {nullptr};
+    PredicateProgram* _prog {nullptr};
 
-    FilterProcessor(PredicateProgram* exprProg);
+    FilterProcessor(PredicateProgram* prog);
     ~FilterProcessor() final = default ;
 };
